feat: migrate to eslint v9
