+++ svelte.mjs
@@ -1,0 +1,62 @@
+import parser from "svelte-eslint-parser";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import js from "@eslint/js";
+import { FlatCompat } from "@eslint/eslintrc";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const compat = new FlatCompat({
+  baseDirectory: __dirname,
+  recommendedConfig: js.configs.recommended,
+  allConfig: js.configs.all,
+});
+
+export default [
+  ...compat.extends("./index.mjs", "plugin:svelte/recommended"),
+  {
+    languageOptions: {
+      ecmaVersion: 5,
+      sourceType: "script",
+
+      parserOptions: {
+        extraFileExtensions: [".svelte"],
+      },
+    },
+
+    rules: {
+      "local-rules/no-svelte-store-in-api": "error",
+    },
+  },
+  {
+    files: ["**/*.svelte"],
+
+    languageOptions: {
+      parser: parser,
+      ecmaVersion: 5,
+      sourceType: "script",
+
+      parserOptions: {
+        parser: "@typescript-eslint/parser",
+      },
+    },
+
+    rules: {
+      "import/order": [
+        "error",
+        {
+          alphabetize: {
+            order: "asc",
+          },
+        },
+      ],
+    },
+  },
+  {
+    files: ["scripts/**/*.mjs", "scripts/**/*.ts"],
+
+    rules: {
+      "no-console": "off",
+    },
+  },
+];
