+++ tsconfig.eslint.json
@@ -1,0 +1,6 @@
+{
+  "extends": ["./tsconfig.json"],
+  "compilerOptions": {
+    "allowJs": true
+  }
+}
